requirements.txt
fastapi
uvicorn[standard]
pillow
python-dotenv
google-generativeai
requests
